requirements.txt
requests>=2.28.0
pdfplumber>=0.7.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
faiss-gpu>=1.7.0
sentence-transformers>=2.2.0
einops>=0.6.0
beautifulsoup4>=4.11.0
pdfplumber>=0.7.0
lxml>=4.9.0
numpy>=1.24.0
tqdm>=4.65.0
huggingface-hub>=0.16.0